Moved sendInterviewInvitationNotifications from services/NotificationsSchedulerService to  eventHandles/InterviewEventHandler
